pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "graphium"
description = "Graphium: Scaling molecular GNNs to infinity."
dynamic = ["version"]
authors = [
    { name = "Dominique Beaini", email = "dominique@valencediscovery.com" },
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "click",
    "loguru",
    "omegaconf >=2.0.0",
    "tqdm",
    "platformdirs",
    # scientific
    "numpy",
    "scipy >=1.4",
    "pandas >=1.0",
    "scikit-learn",
    "fastparquet",
    # viz
    "matplotlib >=3.0.1",
    "seaborn",
    # cloud IO
    "fsspec >=2021.6",
    "s3fs >=2021.6",
    "gcsfs >=2021.6",
    # ML packages
    "lightning >=2.0",
    "torchmetrics >=0.7.0,<0.11",
    "ogb",
    "torch-geometric >=2.0",
    "wandb",
    "mup",
    "torch_sparse >=0.6",
    "torch_cluster >=1.5",
    "torch_scatter >=2.0",
    # chemistry
    "datamol >=0.10",
]

[project.scripts]
graphium = "graphium.cli.main:main_cli"
 graphium-train = "graphium.cli.train_finetune:cli"

[project.urls]
Website = "https://graphium.datamol.io/"
"Source Code" = "https://github.com/datamol-io/graphium"
"Bug Tracker" = "https://github.com/datamol-io/graphium/issues"
Documentation = "https://graphium-docs.datamol.io/"

[tool.setuptools]
include-package-data = true

[tool.setuptools_scm]
fallback_version = "dev"

[tool.setuptools.packages.find]
where = ["."]
include = ["graphium", "graphium.*", "expts", "expts.*"]
exclude = []
namespaces = true

[tool.setuptools.package-data]
"graphium" = ["**/*.csv", "**/*.parquet", "**/*.yaml"]

[tool.black]
line-length = 110
target-version = ['py310', 'py311']
include = '\.pyi?$'

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--verbose --durations=10 -n 1 --cov=graphium --cov-fail-under=60 --cov-report xml --cov-report term"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning:ray.*:",
    "ignore::DeprecationWarning:numba.*:",
    "ignore::DeprecationWarning:lightning_fabric.*:",
    "ignore::DeprecationWarning:pytorch_lightning.*:",
    "ignore::DeprecationWarning:pkg_resources.*:",
]
markers = [
    "ipu: marks tests that are specific to the IPU (deselect with '-m \"not ipu\"')",
]

[tool.coverage.run]
source = ["graphium/"]
disable_warnings = ["no-data-collected"]
data_file = ".coverage/coverage"

[tool.coverage.report]
omit = ["graphium/__init__.py", "graphium/_version.py"]

[tool.coverage.xml]
output = "coverage.xml"
